refactor(eval): move expand and variable resolution to separate package
