+++ internal/eval/resolvers/variables_test.go
@@ -1,8 +1,8 @@
-package eval_test
+package resolvers_test
 
 import (
 	"github.com/b2wdigital/restQL-golang/internal/domain"
-	"github.com/b2wdigital/restQL-golang/internal/eval"
+	"github.com/b2wdigital/restQL-golang/internal/eval/resolvers"
 	"reflect"
 	"testing"
 )
@@ -78,7 +78,7 @@ func TestResolveVariables(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := eval.ResolveVariables(tt.query, tt.input)
+			got := resolvers.ResolveVariables(tt.query, tt.input)
 			if !reflect.DeepEqual(got, tt.expected) {
 				t.Errorf("ResolveVariables = %#+v. Want %#+v", got, tt.expected)
 			}
